+++ common/src/main/java/com/wynntils/models/gear/GearChatEncoding.java
@@ -208,8 +208,8 @@ public class GearChatEncoding {
         // create chat gear stack
         // Note that the chat procotol does not allow for shiny stats to be transferred
         GearInstance gearInstance =
-                GearInstance.create(gearInfo, identifications, powderList, rerolls, Optional.empty(), Optional.empty());
-        return new GearItem(gearInfo, gearInstance);
+                GearInstance.create(gearInfo, identifications, powderList, rerolls, Optional.empty());
+        return new GearItem(gearInfo, gearInstance, Optional.empty(), Optional.empty()); // TODO: populate setInfo but not setInstance
     }
 
     public Matcher gearChatEncodingMatcher(String str) {
