+++ common/src/main/java/com/wynntils/models/gear/GearModel.java
@@ -106,43 +106,9 @@ public final class GearModel extends Model {
             WynntilsMod.warn("Tier for " + gearInfo.name() + " is reported as " + result.tier());
         }
 
-        // Set parsing
-        Optional<SetInstance> setInstance = Optional.empty();
-        if (gearInfo.tier() == GearTier.SET && gearInfo.setInfo().isPresent()) {
-            // We have a set and the information required to make instances
-            // Go through all of user's active gear and figure out how many of this set is active
-            Pair<Integer, Integer> setCount = getActiveSetCount(gearInfo.setInfo().get().name());
-            SetInfo setInfo = gearInfo.setInfo().get();
-
-            // we are wearing some item from this set
-            setInstance = Optional.of(new SetInstance(setInfo, setCount.a(), setCount.b(), activeSetsCache.get(setInfo.name())));
-            if (!activeSetsCache.containsKey(setInfo.name())) {
-                activeSetsCache.put(setInfo.name(), new ArrayList<>());
-            }
-            activeSetsCache.get(setInfo.name()).add(setInstance.get());
-            activeSetsCache.get(setInfo.name()).forEach(si -> {
-                if (si.getWynncraftCount() != setCount.a()) {
-                    si.setWynncraftCount(setCount.a());
-                }
-                if (si.getTrueCount() != setCount.b()) {
-                    si.setTrueCount(setCount.b());
-                }
-                if (si.getSetInstances() != activeSetsCache.get(setInfo.name())) {
-                    si.setSetInstances(activeSetsCache.get(setInfo.name()));
-                }
-                System.out.println("updated SetInstance to " + si);
-            });
-
-            // todo remove
-            if (gearInfo.name().equals("Morph-Gold")) {
-                System.out.println("Found SetInstance for " + gearInfo.name() + ": " + setInstance);
-            } else {
-//                System.out.println("Found SetInstance for " + gearInfo.name() + ": " + setInstance);
-            }
-        }
 
         return GearInstance.create(
-                gearInfo, result.identifications(), result.powders(), result.rerolls(), result.shinyStat(), setInstance);
+                gearInfo, result.identifications(), result.powders(), result.rerolls(), result.shinyStat());
     }
 
     // For parsing gear from the gear viewer
@@ -150,7 +116,7 @@ public final class GearModel extends Model {
         WynnItemParseResult result = WynnItemParser.parseInternalRolls(gearInfo, itemData);
 
         return GearInstance.create(
-                gearInfo, result.identifications(), result.powders(), result.rerolls(), result.shinyStat(), Optional.empty());
+                gearInfo, result.identifications(), result.powders(), result.rerolls(), result.shinyStat());
     }
 
     public CraftedGearItem parseCraftedGearItem(ItemStack itemStack) {
@@ -236,59 +202,5 @@ public final class GearModel extends Model {
 
     public Stream<GearInfo> getAllGearInfos() {
         return gearInfoRegistry.getGearInfoStream();
-    }
-
-    /**
-     * Goes through user's equipped gear (all armour, accessories, held weapon) and returns count of specified
-     * @return <wynncraft count, true count>
-     */
-    private Pair<Integer, Integer> getActiveSetCount(String setName) {
-        Pair<Integer, Integer> count = Pair.of(0, 0);
-
-        for (ItemStack itemStack : McUtils.inventory().armor) {
-            int wynncraftCount = countSet(itemStack, setName);
-            if (wynncraftCount == -1) continue;
-
-            count = Pair.of(wynncraftCount, count.b() + 1);
-        }
-
-        int[] accessorySlots = {9, 10, 11, 12};
-        if (McUtils.player().hasContainerOpen()) {
-            // Scale according to server chest size
-            // Eg. 3 row chest size = 27 (ends on i=26 since 0-index), we would get accessory slots {27, 28, 29, 30}
-            int baseSize = McUtils.player().containerMenu.getItems().size();
-            accessorySlots = new int[]{baseSize, baseSize + 1, baseSize + 2, baseSize + 3};
-        }
-        for (int i : accessorySlots) {
-            ItemStack itemStack = McUtils.inventory().getItem(i);
-            int wynncraftCount = countSet(itemStack, setName);
-            if (wynncraftCount == -1) continue;
-
-            count = Pair.of(wynncraftCount, count.b() + 1);
-        }
-
-        // held item - must check if it's actually valid before counting
-        ItemStack itemInHand = McUtils.player().getItemInHand(InteractionHand.MAIN_HAND);
-        if (InventoryUtils.itemRequirementsMet(itemInHand)) {
-            int wynncraftCount = countSet(itemInHand, setName);
-            if (wynncraftCount == -1) return count;
-
-            count = Pair.of(wynncraftCount, count.b() + 1);
-        }
-
-        return count;
-    }
-
-    /**
-     * @return Wynncraft's count of the set if the set matches the specified name, or -1 if it doesn't
-     */
-    private int countSet(ItemStack itemStack, String setName) {
-        for (StyledText line : LoreUtils.getLore(itemStack)) {
-            Matcher nameMatcher = SET_PATTERN.matcher(line.getString());
-            if (nameMatcher.matches() && nameMatcher.group(1).equals(setName)) {
-                return Integer.parseInt(nameMatcher.group(2));
-            }
-        }
-        return -1;
     }
 }
