+++ common/src/main/java/com/wynntils/models/gear/type/GearInfo.java
@@ -21,8 +21,7 @@ public record GearInfo(
         GearMetaInfo metaInfo,
         GearRequirements requirements,
         FixedStats fixedStats,
-        List<Pair<StatType, StatPossibleValues>> variableStats,
-        Optional<SetInfo> setInfo) {
+        List<Pair<StatType, StatPossibleValues>> variableStats) {
     public StatPossibleValues getPossibleValues(StatType statType) {
         return this.variableStats().stream()
                 .filter(p -> p.key().equals(statType))
