+++ common/src/main/java/com/wynntils/models/items/properties/SetItemProperty.java
@@ -1,0 +1,12 @@
+package com.wynntils.models.items.properties;
+
+import com.wynntils.models.gear.type.SetInfo;
+import com.wynntils.models.gear.type.SetInstance;
+
+import java.util.Optional;
+
+public interface SetItemProperty {
+    Optional<SetInfo> getSetInfo();
+
+    Optional<SetInstance> getSetInstance();
+}
